widget ui: hide public profile checkbox if not manually enabled elsewhere
